Add routers and new update project structure
Separate packages for keyboards, handlers,
and working with the database have been created.

In the keyboard package, two models have been
created for general and registration keyboards.

In the package for handlers, general and
registration handlers have been separated.

The database has been moved to the appropriate
folder, and a module for working with it has been
created there.

.gitignore updated